refactor(web/auth): create LoginForm component with Google and KakaoTalk sign-in options
- Updated the login page to import the new LoginForm component.
- Introduced SignInWithGoogleButton and SignInWithKakaotalkButton components for respective sign-in functionalities.
- Enhanced the user interface with a card layout for a better login experience.
