+++ apps/web/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/app/auth/components/login-form";
+import { LoginForm } from "@/components/login-form";
 
 export default function Page() {
   return (
